bc
UI FIXED

+++ src/app/login/component/sign-up/sign-up.component.ts
@@ -12,21 +12,26 @@ export class SignUpComponent implements OnInit {
 
   email:string="";
   password:string="";
-  orgId:string="";
+  orgId:string="abc";
+  confirmPassword:string='';
   constructor(private loginService:LoginService,private router:Router) { }
 
   ngOnInit() {
   }
 
   registerUser(){
-    let obj:UserRegister={email:this.email,password:this.password,orgId:this.orgId};
-    this.loginService.registerUser(obj).subscribe(v=>{
-      if(v) {
-        localStorage.setItem("token",v.token);
-        this.loginService.setIsLoggedIn(true);
-        this.router.navigate(['/']); 
-       }
-    });
+    
+    if(this.email && this.password && this.confirmPassword && this.password==this.confirmPassword){
+      let obj:UserRegister={email:this.email,password:this.password,orgId:this.orgId};
+      this.loginService.registerUser(obj).subscribe(v=>{
+        if(v) {
+          localStorage.setItem("token",v.token);
+          this.loginService.setIsLoggedIn(true);
+          this.router.navigate(['/']); 
+         }
+      });
+    }
+    
   }
 
 }
